Split claim extension methods by responsibility
Separated IList<Claim> and ClaimsPrincipal extension methods into ClaimsExtensions.cs and ClaimsPrincipalExtensions.cs, respectively. Improved XML documentation and type safety, enhancing code clarity and maintainability.
