+++ src/TinyHelpers.AspNetCore/Extensions/NavigationManagerExtensions.cs
@@ -4,8 +4,19 @@ using Microsoft.AspNetCore.WebUtilities;
 
 namespace TinyHelpers.AspNetCore.Extensions;
 
+/// <summary>
+/// Adds query-string parsing helpers that keep component code focused on behavior instead of URI manipulation.
+/// </summary>
 public static class NavigationManagerExtensions
 {
+    /// <summary>
+    /// Tries to read a typed value from the current URI query string without requiring callers to parse the raw string themselves.
+    /// </summary>
+    /// <typeparam name="T">The expected query-string value type.</typeparam>
+    /// <param name="navManager">The navigation manager that exposes the current URI.</param>
+    /// <param name="key">The query-string key to match.</param>
+    /// <param name="value">When this method returns, contains the converted value if the key exists and can be parsed.</param>
+    /// <returns><see langword="true" /> when the query string contains a parsable value for <paramref name="key" />; otherwise, <see langword="false" />.</returns>
     public static bool TryGetQueryString<T>(this NavigationManager navManager, string key, [NotNullWhen(true)] out T? value)
     {
         var uri = navManager.ToAbsoluteUri(navManager.Uri);
