+++ src/TinyHelpers.AspNetCore/Middlewares/ApplicationBuilderExtensions.cs
@@ -2,8 +2,19 @@
 
 namespace TinyHelpers.AspNetCore.Middlewares;
 
+/// <summary>
+/// Provides middleware registration helpers.
+/// </summary>
 public static class ApplicationBuilderExtensions
 {
-    public static IApplicationBuilder UseRequestRewind(this IApplicationBuilder app)
-        => app.UseMiddleware<EnableRequestRewindMiddleware>();
+    extension(IApplicationBuilder app)
+    {
+        /// <summary>
+        /// Enables request buffering so downstream middleware and services can re-read the body after an earlier component has inspected it.
+        /// </summary>
+        /// <param name="app">The application pipeline being configured.</param>
+        /// <returns>The <see cref="IApplicationBuilder" /> builder instance.</returns>
+        public IApplicationBuilder UseRequestRewind()
+            => app.UseMiddleware<EnableRequestRewindMiddleware>();
+    }
 }
