+++ src/TinyHelpers.AspNetCore/OpenApi/OpenApiSchemaHelper.cs
@@ -5,8 +5,16 @@ using Microsoft.OpenApi.Models;
 
 namespace TinyHelpers.AspNetCore.OpenApi;
 
+/// <summary>
+/// Creates reusable schema fragments so OpenAPI transformers can express the same contract details without duplicating schema setup.
+/// </summary>
 public static class OpenApiSchemaHelper
 {
+    /// <summary>
+    /// Creates a plain string schema with an optional default value for cases where the document needs to stay minimal.
+    /// </summary>
+    /// <param name="defaultValue">The optional default value to expose in the schema.</param>
+    /// <returns>A string schema.</returns>
     public static OpenApiSchema CreateStringSchema(string? defaultValue = null)
     {
         var schema = new OpenApiSchema
@@ -18,6 +26,13 @@ public static class OpenApiSchemaHelper
         return schema;
     }
 
+    /// <summary>
+    /// Creates a schema with the specified type and format when the caller already knows the wire representation.
+    /// </summary>
+    /// <typeparam name="TValue">The CLR type associated with the schema.</typeparam>
+    /// <param name="type">The OpenAPI type name.</param>
+    /// <param name="format">The optional format string.</param>
+    /// <returns>A schema configured with the supplied type metadata.</returns>
     public static OpenApiSchema CreateSchema<TValue>(string type, string? format = null)
     {
         var schema = new OpenApiSchema
@@ -29,6 +44,14 @@ public static class OpenApiSchemaHelper
         return schema;
     }
 
+    /// <summary>
+    /// Creates a typed schema with a default value so generated clients can reflect the common case.
+    /// </summary>
+    /// <typeparam name="TValue">The CLR value type associated with the schema.</typeparam>
+    /// <param name="type">The OpenAPI type name.</param>
+    /// <param name="format">The optional format string.</param>
+    /// <param name="defaultValue">The optional default value.</param>
+    /// <returns>A schema configured with the supplied type metadata and default.</returns>
     public static OpenApiSchema CreateSchema<TValue>(string type, string? format, TValue? defaultValue = null) where TValue : struct
     {
         var schema = new OpenApiSchema
@@ -41,6 +64,12 @@ public static class OpenApiSchemaHelper
         return schema;
     }
 
+    /// <summary>
+    /// Creates a string enum schema from a known list of values so callers do not need to hand-build the enum metadata.
+    /// </summary>
+    /// <param name="values">The allowed values to expose in the schema.</param>
+    /// <param name="defaultValue">The optional default value.</param>
+    /// <returns>A string schema with the supplied enum values.</returns>
     public static OpenApiSchema CreateSchema(IEnumerable<string> values, string? defaultValue = null)
     {
         var schema = new OpenApiSchema
@@ -53,6 +82,12 @@ public static class OpenApiSchemaHelper
         return schema;
     }
 
+    /// <summary>
+    /// Creates an enum schema from a CLR enumeration so the generated document stays in sync with the source type.
+    /// </summary>
+    /// <typeparam name="TEnum">The enumeration type to describe.</typeparam>
+    /// <param name="defaultValue">The optional default enum value.</param>
+    /// <returns>A string enum schema populated with every declared member.</returns>
     public static OpenApiSchema CreateSchema<TEnum>(TEnum? defaultValue = null) where TEnum : struct, Enum
     {
         var schema = new OpenApiSchema
@@ -73,8 +108,16 @@ using Microsoft.OpenApi;
 
 namespace TinyHelpers.AspNetCore.OpenApi;
 
+/// <summary>
+/// Creates reusable schema fragments so OpenAPI transformers can express the same contract details without duplicating schema setup.
+/// </summary>
 public static class OpenApiSchemaHelper
 {
+    /// <summary>
+    /// Creates a plain string schema with an optional default value for cases where the document needs to stay minimal.
+    /// </summary>
+    /// <param name="defaultValue">The optional default value to expose in the schema.</param>
+    /// <returns>A string schema.</returns>
     public static OpenApiSchema CreateStringSchema(string? defaultValue = null)
     {
         var schema = new OpenApiSchema
@@ -86,6 +129,13 @@ public static class OpenApiSchemaHelper
         return schema;
     }
 
+    /// <summary>
+    /// Creates a schema with the specified type and format when the caller already knows the wire representation.
+    /// </summary>
+    /// <typeparam name="TValue">The CLR type associated with the schema.</typeparam>
+    /// <param name="type">The OpenAPI type value.</param>
+    /// <param name="format">The optional format string.</param>
+    /// <returns>A schema configured with the supplied type metadata.</returns>
     public static OpenApiSchema CreateSchema<TValue>(JsonSchemaType type, string? format = null)
     {
         var schema = new OpenApiSchema
@@ -97,6 +147,14 @@ public static class OpenApiSchemaHelper
         return schema;
     }
 
+    /// <summary>
+    /// Creates a typed schema with a default value so generated clients can reflect the common case.
+    /// </summary>
+    /// <typeparam name="TValue">The CLR value type associated with the schema.</typeparam>
+    /// <param name="type">The OpenAPI type value.</param>
+    /// <param name="format">The optional format string.</param>
+    /// <param name="defaultValue">The optional default value.</param>
+    /// <returns>A schema configured with the supplied type metadata and default.</returns>
     public static OpenApiSchema CreateSchema<TValue>(JsonSchemaType type, string? format, TValue? defaultValue = null) where TValue : struct
     {
         var schema = new OpenApiSchema
@@ -109,6 +167,12 @@ public static class OpenApiSchemaHelper
         return schema;
     }
 
+    /// <summary>
+    /// Creates a string enum schema from a known list of values so callers do not need to hand-build the enum metadata.
+    /// </summary>
+    /// <param name="values">The allowed values to expose in the schema.</param>
+    /// <param name="defaultValue">The optional default value.</param>
+    /// <returns>A string schema with the supplied enum values.</returns>
     public static OpenApiSchema CreateSchema(IEnumerable<string> values, string? defaultValue = null)
     {
         var schema = new OpenApiSchema
@@ -121,6 +185,12 @@ public static class OpenApiSchemaHelper
         return schema;
     }
 
+    /// <summary>
+    /// Creates an enum schema from a CLR enumeration so the generated document stays in sync with the source type.
+    /// </summary>
+    /// <typeparam name="TEnum">The enumeration type to describe.</typeparam>
+    /// <param name="defaultValue">The optional default enum value.</param>
+    /// <returns>A string enum schema populated with every declared member.</returns>
     public static OpenApiSchema CreateSchema<TEnum>(TEnum? defaultValue = null) where TEnum : struct, Enum
     {
         var schema = new OpenApiSchema
