+++ src/TinyHelpers.AspNetCore/OpenApi/Transformers/OpenApiParametersOperationTransformer.cs
@@ -7,6 +7,13 @@ namespace TinyHelpers.AspNetCore.OpenApi;
 
 internal class OpenApiParametersOperationFilter(OpenApiOperationOptions options) : IOpenApiOperationTransformer
 {
+    /// <summary>
+    /// Copies custom parameter metadata into an operation so shared rules can be registered once and reused across endpoints.
+    /// </summary>
+    /// <param name="operation">The OpenAPI operation being enriched.</param>
+    /// <param name="context">The operation-generation context.</param>
+    /// <param name="cancellationToken">A token that signals request cancellation.</param>
+    /// <returns>A task that completes when the parameters have been copied.</returns>
     public Task TransformAsync(OpenApiOperation operation, OpenApiOperationTransformerContext context, CancellationToken cancellationToken)
     {
         if (options.Parameters.Count > 0)
@@ -32,6 +39,13 @@ namespace TinyHelpers.AspNetCore.OpenApi;
 
 internal class OpenApiParametersOperationFilter(OpenApiOperationOptions options) : IOpenApiOperationTransformer
 {
+    /// <summary>
+    /// Copies custom parameter metadata into an operation so shared rules can be registered once and reused across endpoints.
+    /// </summary>
+    /// <param name="operation">The OpenAPI operation being enriched.</param>
+    /// <param name="context">The operation-generation context.</param>
+    /// <param name="cancellationToken">A token that signals request cancellation.</param>
+    /// <returns>A task that completes when the parameters have been copied.</returns>
     public Task TransformAsync(OpenApiOperation operation, OpenApiOperationTransformerContext context, CancellationToken cancellationToken)
     {
         if (options.Parameters.Count > 0)
