+++ src/TinyHelpers.AspNetCore/OpenApi/Transformers/StrictNumericSchemaTransformer.cs
@@ -7,6 +7,13 @@ namespace TinyHelpers.AspNetCore.OpenApi.Transformers;
 
 internal sealed class StrictNumericSchemaTransformer : IOpenApiSchemaTransformer
 {
+    /// <summary>
+    /// Removes string fallbacks from numeric schemas so clients interpret the contract as strictly numeric.
+    /// </summary>
+    /// <param name="schema">The schema being enriched.</param>
+    /// <param name="context">The schema-generation context.</param>
+    /// <param name="cancellationToken">A token that signals request cancellation.</param>
+    /// <returns>A task that completes when the schema has been updated.</returns>
     public Task TransformAsync(OpenApiSchema schema, OpenApiSchemaTransformerContext context, CancellationToken cancellationToken)
     {
         if (schema.Type is null)
